# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Friction MVP 1.0.0 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stored-sentences:
    get:
      operationId: listStoredSentences
      tags: [stored-sentences]
      summary: List stored sentences for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: favorite
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Sentences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoredSentence"
    post:
      operationId: createStoredSentence
      tags: [stored-sentences]
      summary: Store a sentence from an article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoredSentenceBody"
      responses:
        "201":
          description: Created sentence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredSentence"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stored-sentences/{id}:
    delete:
      operationId: deleteStoredSentence
      tags: [stored-sentences]
      summary: Delete a stored sentence
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stored-sentences/{id}/favorite:
    patch:
      operationId: toggleStoredSentenceFavorite
      tags: [stored-sentences]
      summary: Toggle favorite status
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isFavorite:
                  type: boolean
              required:
                - isFavorite
      responses:
        "200":
          description: Updated sentence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoredSentence"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Reading ──────────────────────────────────────
  /reading-records:
    get:
      operationId: getReadingRecord
      tags: [reading]
      summary: Get reading progress for a user+article
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: articleId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reading record (or null if none)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingRecordNullable"
    put:
      operationId: upsertReadingRecord
      tags: [reading]
      summary: Create or update reading progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertReadingRecordBody"
      responses:
        "200":
          description: Upserted reading record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user-article-reads:
    post:
      operationId: createUserArticleRead
      tags: [reading]
      summary: Record article completion (完독)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserArticleReadBody"
      responses:
        "201":
          description: Created read record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserArticleRead"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user-article-reads/check:
    get:
      operationId: checkArticleRead
      tags: [reading]
      summary: Check if user has completed reading an article
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: articleId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Read status
          content:
            application/json:
              schema:
                type: object
                properties:
                  isRead:
                    type: boolean
                  completedAt:
                    type: string
                    format: date-time
                    nullable: true
                required:
                  - isRead

  # ─── TeamCollections ──────────────────────────────
...
  /send-records:
    get:
      operationId: listSendRecords
      tags: [send-records]
      summary: List send records for a user
      parameters:
        - name: senderId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Send records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SendRecordWithDetails"
    post:
      operationId: sendArticle
      tags: [send-records]
      summary: Send an article to a recipient
      description: "Creates SendRecord + Inbox entry. Auto-assigns 06:00 or 18:00 KST delivery slot. Article must be LETTER status."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendArticleBody"
      responses:
        "201":
          description: Send record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendRecordWithDetails"
        "400":
          description: Article not LETTER or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, nickname, createdAt, updatedAt]
...
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [DRAFT, DIVIDING, CLOSING, LETTER]
        pages:
          type: array
          items:
            type: string
          nullable: true
        style:
          type: object
          nullable: true
        letterAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, authorId, title, content, status, createdAt, updatedAt]
...
    StoredSentence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        text:
          type: string
        position:
          type: object
          nullable: true
        isFavorite:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, articleId, text, isFavorite, createdAt]
    CreateStoredSentenceBody:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        text:
          type: string
          minLength: 1
        position:
          type: object
          nullable: true
      required: [userId, articleId, text]

    # ─── Reading ────────────────────────────────────
    ReadingRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        currentPage:
          type: integer
        scrollPosition:
          type: number
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, articleId, currentPage, scrollPosition, updatedAt]
    ReadingRecordNullable:
      type: object
      properties:
        record:
          nullable: true
          oneOf:
            - $ref: "#/components/schemas/ReadingRecord"
            - type: "null"
      required: [record]
    UpsertReadingRecordBody:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        currentPage:
          type: integer
          minimum: 0
        scrollPosition:
          type: number
          minimum: 0
          maximum: 1
      required: [userId, articleId, currentPage, scrollPosition]
    UserArticleRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        completedAt:
          type: string
          format: date-time
      required: [id, userId, articleId, completedAt]
    CreateUserArticleReadBody:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
      required: [userId, articleId]

    # ─── TeamCollection ─────────────────────────────
...
    SendRecordWithDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        recipientId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        inboxId:
          type: string
          format: uuid
        deliverySlot:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
        article:
          $ref: "#/components/schemas/Article"
        recipient:
          $ref: "#/components/schemas/User"
      required: [id, senderId, recipientId, articleId, inboxId, deliverySlot, sentAt]
    SendArticleBody:
      type: object
      properties:
        senderId:
          type: string
          format: uuid
        recipientId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
      required: [senderId, recipientId, articleId]