# lib/api-zod/src/generated/types/createStoredSentenceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Friction MVP 1.0.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateStoredSentenceBodyPosition } from "./createStoredSentenceBodyPosition";

export interface CreateStoredSentenceBody {
  userId: string;
  articleId: string;
  /** @minLength 1 */
  text: string;
  position?: CreateStoredSentenceBodyPosition;
}