# lib/api-zod/src/generated/types/sendRecordWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Friction MVP 1.0.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";
import type { User } from "./user";

export interface SendRecordWithDetails {
  id: string;
  senderId: string;
  recipientId: string;
  articleId: string;
  inboxId: string;
  deliverySlot: Date;
  sentAt: Date;
  article?: Article;
  recipient?: User;
}