# lib/api-zod/src/generated/types/userArticleRead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Friction MVP 1.0.0 API
 * OpenAPI spec version: 0.1.0
 */

export interface UserArticleRead {
  id: string;
  userId: string;
  articleId: string;
  completedAt: Date;
}